+++ packages/examples/resources/python/hello.py
@@ -1,3 +1,4 @@
+# if hello2 is not resolved the file is not availale on the language server
 from hello2 import print_hello
 
 print_hello()