+++ packages/examples/src/common/client/utils.ts
@@ -3,43 +3,15 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { Uri } from 'vscode';
-import type { Logger } from 'monaco-languageclient/tools';
-import { useWorkerFactory, type WorkerLoader } from 'monaco-languageclient/workerFactory';
 import { RegisteredMemoryFile } from '@codingame/monaco-vscode-files-service-override';
 import type { IStoredWorkspace } from '@codingame/monaco-vscode-configuration-service-override';
+import type { Uri } from 'vscode';
 
 export const disableElement = (id: string, disabled: boolean) => {
     const button = document.getElementById(id) as HTMLButtonElement | HTMLInputElement | null;
     if (button !== null) {
         button.disabled = disabled;
     }
-};
-
-export const defineDefaultWorkerLoaders: () => Record<string, WorkerLoader> = () => {
-    return {
-        // if you import monaco api as 'monaco-editor': monaco-editor/esm/vs/editor/editor.worker.js
-        TextEditorWorker: () => new Worker(
-            new URL('@codingame/monaco-vscode-editor-api/esm/vs/editor/editor.worker.js', import.meta.url),
-            { type: 'module' }
-        ),
-        TextMateWorker: () => new Worker(
-            new URL('@codingame/monaco-vscode-textmate-service-override/worker', import.meta.url),
-            { type: 'module' }
-        ),
-        // these are other possible workers not configured by default
-        OutputLinkDetectionWorker: undefined,
-        LanguageDetectionWorker: undefined,
-        NotebookEditorWorker: undefined,
-        LocalFileSearchWorker: undefined
-    };
-};
-
-export const configureMonacoWorkers = (logger?: Logger) => {
-    useWorkerFactory({
-        workerLoaders: defineDefaultWorkerLoaders(),
-        logger
-    });
 };
 
 export const createDefaultWorkspaceFile = (workspaceFile: Uri, workspacePath: string) => {