+++ packages/wrapper-react/test/helper.ts
@@ -5,15 +5,7 @@
 
 import { LogLevel } from '@codingame/monaco-vscode-api';
 import type { WrapperConfig } from 'monaco-editor-wrapper';
-import { useWorkerFactory } from 'monaco-languageclient/workerFactory';
-
-export const configureMonacoWorkers = () => {
-    useWorkerFactory({
-        workerLoaders: {
-            TextEditorWorker: () => new Worker(new URL('@codingame/monaco-vscode-editor-api/esm/vs/editor/editor.worker.js', import.meta.url), { type: 'module' }),
-        }
-    });
-};
+import { configureDefaultWorkerFactory } from 'monaco-editor-wrapper/workers/workerLoaders';
 
 export const createDefaultWrapperConfig = (): WrapperConfig => {
     return {
@@ -29,7 +21,7 @@ export const createDefaultWrapperConfig = (): WrapperConfig => {
                     fileExt: 'js'
                 }
             },
-            monacoWorkerFactory: configureMonacoWorkers
+            monacoWorkerFactory: configureDefaultWorkerFactory
         }
     };
 };