+++ verify/vite/vite.config.ts
@@ -7,7 +7,7 @@ import { defineConfig } from 'vite';
 
 export default defineConfig({
     preview: {
-        port: 8082
+        port: 8083
     },
     worker: {
         format: 'es'