+++ packages/client/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this npm module are documented in this file.
 
+## [9.4.0] - 2025-02-18
+
+- Handle languageclient errors cases more robustly [#859](https://github.com/TypeFox/monaco-languageclient/pull/859)
+- Updated all `@codingame/monaco-vscode` packages to `14.0.4`.
+
 ## [9.3.0] - 2025-02-12
 
 - Updated all `@codingame/monaco-vscode` packages to `14.0.2`.